requirements.txt
-i https://pypi.org/simple
certifi==2020.11.8
chardet==3.0.4
idna==2.10
pyjwt==1.7.1
requests==2.25.0
urllib3==1.26.2
